added updates and fixes

+++ MemberData/MemberServerProfile.cs
@@ -4,8 +4,8 @@ namespace Motion.Core.Data.MemberData
 	public class MemberServerProfile
 	{
 		public int MemberDeviceID { get; set; }
-		public long UpdateFlag { get; set; }
-		public long UpdateFlagChanges { get; set; }
+		public int UpdateFlag { get; set; }
+		public int UpdateFlagChanges { get; set; }
 		public int SynchronizationID { get; set; }
 		public string ServerDateTime { get; set; }
